session: avoid wrapping redirects in exn handlers
When redirecting from http -> https and the https request would
fail (for example, with an SSL error), the exception would propagate
up to the exception handler installed on the http request, resulting
in a confusing "released resource was not leased" error, because the
http connection would end up being released twice.

Also, move contracts to provide form.