response: limit how long a response can take to close
Previously, closing a response would wait on the response data port to
get drained, but that meant a buggy (or adversarial) server could
block the close perpetually. Likewise, draining a connection during
redirect or when #:close? was #t had the same issue.

This change reworks closing and draining to have timeouts and to
behave as reasonably as they can in the face of such responses, by
still draining the data for compliant servers and timing out after one
second on close and after the response timeout config on drain.